requirements.txt
asgiref==3.7.2
boto3==1.26.83
botocore==1.29.165
Django==5.0.1
django-cors-headers==4.3.1
django-environ==0.11.2
django-multiselectfield==0.1.12
django-storages==1.14.2
djangorestframework==3.14.0
gunicorn==21.2.0
jmespath==1.0.1
packaging==23.2
Pillow==10.1.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
s3transfer==0.6.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==1.26.18
whitenoise==5.3.0
uvicorn==0.25.0
flake8==7.0.0
